GLT-1412: refactored and improved SampleType conversion
